Update BinarySearchTreeWithDups.java
Everything is currently all working on my end. I put the efficiency output below.

If running ProjectCPartBDriver.java make sure to comment out the three "// EFFICIENCY VARIABLE PRINT"

```
------------------------------EVALUATING METHOD EFFICIENCY------------------------------
This method is designed to help you figure out if you are fully taking advantage of the sorted nature of a BST.
Essentially what it does is create very lopsided trees and then invokes methods with targets that are "near the root" or "deep in the tree."
If you are fully taking advantage of the sorted nature of the tree, then the method should be faster when the target is near the root.

*****IMPORTANT*****: In order to use these tests, you must add code to the method you are evaluating.
The code should count how many iterations the method took- loops or recursion.
Please see the provided "example code snippet" file in the provided files for examples of what code to add.

Each test below will show two counts.
If the two counts are very different: you might be okay for that method.
If the two iteration counts are equal or very close to each other... 
ALERT!!! You might not be fully taking advantage of the sorted nature of the BST for the  method specified.

Don't forget to remove the test code before submission! :)

Test 1: countEntriesNonRecursive Test A. The two numbers should be very different.
1
5000

Test 2: countEntriesNonRecursive Test B. The two numbers should be very different.
5000
1

Test 3: countGreaterIterative. The two numbers should be very different.
1
5000

Test 4: countGreaterRecursive. The two numbers should be very different.
2
10001

```